pyproject.toml
[tool.poetry]
name = "chatgpt_meraki_analyzer"
version = "0.1.1"
description = "A chatGPT powered Meraki analyzer"
authors = ["John Capobianco <ptcapo@gmail.com>"]
packages = [{ include = "chatgpt_meraki_analyzer", from = "src"}]
readme = "README.md"

[tool.poetry.scripts]
chatgpt_meraki_analyzer = "chatgpt_meraki_analyzer.script:run"

[tool.poetry.dependencies]
python = "^3.8"
rich-click = "^1.6.1"
meraki = "^1.27.0"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
requests = "^2.28.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
